% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{arrange_button}
\alias{arrange_button}
\title{Buttons for \strong{RMarkdown} document}
\usage{
arrange_button(
  label,
  by,
  desc = FALSE,
  status = "default",
  icon = NULL,
  width = NULL,
  ...
)
}
\arguments{
\item{label}{The contents of the button.}

\item{by}{Key(s) defined in \code{shuffle_card} to sort elements.}

\item{desc}{Logical, set to \code{TRUE} to sort in decreasing order.}

\item{status}{Add a class to the buttons, you can use Bootstrap status like 'info',
'primary', 'danger', 'warning' or 'success'. Or use an arbitrary strings to add a
custom class, e.g. : with \code{status = 'myClass'}, buttons will have class \code{btn-myClass}.}

\item{icon}{An optional icon to appear on the button.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'}.}

\item{...}{Named attributes to be applied to the button.}
}
\description{
Arrange a Shuffle grid in \strong{RMarkdown}.
}
\note{
Use \code{\link{rmd_group_buttons}} to position several buttons.
}
