% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{as_svg}
\alias{as_svg}
\title{Convert a `ggplot2`  object to `svg`}
\usage{
as_svg(p, width = 400, height = 300)
}
\arguments{
\item{p}{A `ggplot2` object.}

\item{width}{Width, in numeric and pixels.}

\item{height}{Height, in numeric and pixels.}
}
\description{
\code{\link{shuffle_widget}} generates HTML tags, to represent `ggplot2` objects you need to convert them.
This is the purpose of this function, you can also use `ggiraph` or `plotly`.
}
