% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{filter_cards}
\alias{filter_cards}
\title{Filter a Shuffle grid layout with custom values}
\usage{
filter_cards(session, shuffleId, filter_by, filter_list)
}
\arguments{
\item{session}{The \code{session} object passed to function given to shinyServer.}

\item{shuffleId}{The id of the shuffle container.}

\item{filter_by}{Key defined in \code{shuffle_card} to filter elements, can be \code{"id"}
to refer to the unique ID associated with the card.}

\item{filter_list}{A named list, where names match the names used in \code{filter_by}
for \code{shuffle_card} and value are \code{TRUE} or \code{FALSE}. Alternatively a \code{data.frame}
where first columns contains names and second \code{TRUE} or \code{FALSE}.}
}
\description{
Filter a Shuffle grid layout with custom values
}
\examples{
# TODO
}
