% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{filter_cards_groups}
\alias{filter_cards_groups}
\title{Filter a Shuffle grid layout by groups}
\usage{
filter_cards_groups(shuffleId, groups,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{shuffleId}{The id of the shuffle container.}

\item{groups}{Groups (defined in \code{shuffle_card}) you want to display.}

\item{session}{The \code{session} object passed to function given to shinyServer.}
}
\description{
Use in a \strong{Shiny} app to filter cards.
}
\examples{
if (interactive()) {
  library(ggplot2)
  library(shiny)
  library(shufflecards)

  # utility fun
  is.even <- function(x) x \%\% 2 == 0

  ui <- fluidPage(
    tags$h2("Filter a Shuffle Grid By Groups"),
    fluidRow(
      column(
        width = 3,
        radioButtons(
          inputId = "type_num",
          label = "Odd or Even",
          choices = c("All", "Odd", "Even")
        )
      ),
      column(
        width = 9,
        shuffle_container(
          shuffleId = "gridNum",
          card_list = lapply(
            X = 1:12,
            FUN = function(i) {
              shuffle_card(
                groups = ifelse(is.even(i), "even", "odd"),
                plotOutput(
                  outputId = paste0("plot", i),
                  width = "250px",
                  height = "250px"
                )
              )
            }
          )
        )
      )
    )
  )

  server <- function(input, output, session) {

    # Make individual plots ----
    lapply(
      X = 1:12,
      FUN =  function(i) {
        output[[paste0("plot", i)]] <- renderPlot({
          ggplot() + geom_text(aes(1, 1, label = i), size = 50)
        })
      }
    )

    # Filters ----
    observe({
      if (input$type_num == "All") {
        type_num <- c("even", "odd")
      } else {
        type_num <- tolower(input$type_num)
      }

      # Pass a vector of groups to display
      filter_cards_groups("gridNum", type_num)
    })
  }

  shinyApp(ui, server)
}

}
