% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{md_shuffle_update}
\alias{md_shuffle_update}
\title{Update Shuffle Layout in \strong{RMarkdown} document}
\usage{
md_shuffle_update(shuffleId, selector)
}
\arguments{
\item{shuffleId}{The id of the shuffle container.}

\item{selector}{Selector to trigger update, typically a button,
you can use \href{https://selectorgadget.com/}{SelectorGadget} to find the proper selector.}
}
\description{
Use this if you have Shuffle layout in hidden tab panels in a markdown document.
}
