% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playing-cards.R
\name{playing_card}
\alias{playing_card}
\title{Playing cards}
\usage{
playing_card(suit = c("hearts", "spades", "diamonds", "clubs"),
  value = c("a", 2:10, "j", "q", "k"))
}
\arguments{
\item{suit}{The name of the suit (e.g. \code{"hearts"}).}

\item{value}{The value (e.g. \code{"9"} for a nine).}
}
\description{
Playing cards
}
\examples{
if (interactive()) {
  library(shiny)
  library(shufflecards)

  ui <- fluidPage(
    tags$h3("Playing cards"),
    fluidRow(
      column(
        width = 3,
        "King of Heart",
        playing_card("heart", "k")
      ),
      column(
        width = 3,
        "Random:",
        uiOutput("random"),
        actionButton("draw", "Draw")
      )
    )
  )

  server <- function(input, output, session) {

    output$random <- renderUI({
      input$draw
      playing_card(
        sample(c("hearts", "spades", "diamonds", "clubs"), 1),
        sample(c("a", 2:10, "j", "q", "k"), 1)
      )
    })

  }

  shinyApp(ui, server)
}
}
