% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adding.R
\name{remove_card}
\alias{remove_card}
\title{Remove a card in a Shuffle grid}
\usage{
remove_card(session, shuffleId, cardId)
}
\arguments{
\item{session}{The \code{session} object passed to function given to shinyServer.}

\item{shuffleId}{The id of the shuffle container.}

\item{cardId}{The id of the card to remove.}
}
\description{
Remove a card in a Shuffle grid
}
\examples{
if (interactive()) {
  library(shiny)
  library(shufflecards)

  ui <- fluidPage(
    tags$h2("Remove a card"),
    selectInput(
      inputId = "select",
      label = "Select card to remove",
      choices = paste0("card", 1:4)
    ),
    actionButton("remove", "Remove it!"),
    shuffle_container(
      shuffleId = "grid",
      shuffle_card(
        id = "card1",
        tags$div("My first card", style = "text-align: center; line-height: 200px"),
        style = "border: 2px solid red; border-radius: 5px;",
        width = "300px", # better with fixed width/height
        height = "200px"
      ),
      shuffle_card(
        id = "card2",
        tags$div("Second one", style = "text-align: center; line-height: 200px"),
        style = "border: 2px solid red; border-radius: 5px;",
        width = "300px", # better with fixed width/height
        height = "200px"
      ),
      shuffle_card(
        id = "card3",
        tags$div("Third one", style = "text-align: center; line-height: 200px"),
        style = "border: 2px solid red; border-radius: 5px;",
        width = "300px", # better with fixed width/height
        height = "200px"
      ),
      shuffle_card(
        id = "card4",
        tags$div("Fourth one", style = "text-align: center; line-height: 200px"),
        style = "border: 2px solid red; border-radius: 5px;",
        width = "300px", # better with fixed width/height
        height = "200px"
      )
    )
  )

  server <- function(input, output, session) {
    observeEvent(input$remove, {
      remove_card(session, "grid", input$select)
    }, ignoreInit = TRUE)
  }

  shinyApp(ui, server)
}
}
