% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{rmd_group_buttons}
\alias{rmd_group_buttons}
\title{Group of buttons for Markdown documents}
\usage{
rmd_group_buttons(shuffleId, ..., label = NULL)
}
\arguments{
\item{shuffleId}{The id of the shuffle container.}

\item{...}{\code{\link{arrange_button}} buttons to include.}

\item{label}{Optionnal label to display above buttons.}
}
\description{
Group of buttons for Markdown documents
}
\examples{
# TODO
}
