% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{run_example}
\alias{run_example}
\title{Run examples of use of 'shufflecards'}
\usage{
run_example(example = c("gapminder", "bs-grid", "grid-plot", "filter-sort"))
}
\arguments{
\item{example}{Name of the example, between : \code{"gapminder"}, \code{"bs-grid"}, \code{"grid-plot"}.}
}
\description{
Run examples of use of 'shufflecards'
}
\examples{
if (interactive()) {

run_example("gapminder")

}
}
