% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{shuffle_card}
\alias{shuffle_card}
\title{Shuffle card element}
\usage{
shuffle_card(..., groups = NULL, id = NULL, class = NULL,
  style = NULL, width = NULL)
}
\arguments{
\item{...}{UI elements to include within the card.}

\item{groups}{Character vector of groups used to filtering.}

\item{id}{Cards's id.}

\item{class}{CSS class(es) to apply on the card.}

\item{style}{Inline CSS to apply on the card.}

\item{width}{The width of the container, e.g. \code{'400px'}, or \code{'100\%'}; see \code{\link[htmltools]{validateCssUnit}}.}
}
\description{
Shuffle card element
}
\examples{

# TODO
}
