% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{shuffle_container}
\alias{shuffle_container}
\title{Create a Shuffle container}
\usage{
shuffle_container(shuffleId, ..., card_list = NULL, no_card = NULL,
  width = NULL)
}
\arguments{
\item{shuffleId}{Shuffle's id.}

\item{...}{List of \code{shuffle_card}s to include.}

\item{card_list}{Alternative list of \code{shuffle_card}s to include.}

\item{no_card}{UI definition (or text) to display when all cards are filtered out.}

\item{width}{The width of the container, e.g. \code{'400px'}, or \code{'100\%'}; see \code{\link[htmltools]{validateCssUnit}}.}
}
\description{
Create a Shuffle container
}
\examples{

# TODO
}
