% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_widget.R
\name{shuffle_widget}
\alias{shuffle_widget}
\title{Htmlwidget for 'Shuffle.js'}
\usage{
shuffle_widget(
  ...,
  card_list = NULL,
  shared_data = NULL,
  options = shuffle_options(),
  no_card = NULL,
  use_bs_grid = FALSE,
  width = NULL,
  height = NULL,
  shuffleId = NULL
)
}
\arguments{
\item{...}{List of \code{shuffle_card}s to include.}

\item{card_list}{Alternative list of \code{shuffle_card}s to include.}

\item{shared_data}{A `crosstalk` \code{\link[crosstalk]{SharedData}} object
where a row represent a \code{\link{shuffle_card}} and can be used for filter the grid.}

\item{options}{Options for Shuffle, see \code{\link{shuffle_options}}.}

\item{no_card}{UI definition (or text) to display when all cards are filtered out.}

\item{use_bs_grid}{Allow to use Bootstrap grid system, e.g. same
as in \code{\link[shiny]{column}}. If \code{TRUE}, you can use
a width between 1 and 12 in \code{\link{shuffle_card}}.}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{shuffleId}{Use an explicit element ID for the widget.
Needed to associate the widget to \code{\link{rmd_group_buttons}}.}
}
\description{
Create a grid layout in \strong{RMarkdown}.
 Arrange the grid with \code{\link{arrange_button}}s and filter it with crosstalk inputs.
}
