% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{sort-cards}
\alias{sort-cards}
\alias{sort_cards}
\alias{randomize_cards}
\title{Sort cards in a Shuffle grid layout}
\usage{
sort_cards(session, shuffleId, sort_by, numeric = FALSE,
  decreasing = FALSE)

randomize_cards(session, shuffleId)
}
\arguments{
\item{session}{The \code{session} object passed to function given to shinyServer.}

\item{shuffleId}{The id of the shuffle container.}

\item{sort_by}{Key(s) defined in \code{shuffle_card} to sort elements.}

\item{numeric}{Logical, set to \code{TRUE} if key is a numeric value.}

\item{decreasing}{Logical, set to \code{TRUE} to sort in decreasing order.}
}
\description{
Sort cards in a Shuffle grid layout
}
\examples{

# TODO
}
