% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumbnail.R
\name{thumbnail}
\alias{thumbnail}
\title{Bootstrap thumbnail}
\usage{
thumbnail(title, media, content, href = NULL)
}
\arguments{
\item{title}{Title of the thumbnail.}

\item{media}{A HTML tag img for displaying an image.}

\item{content}{Description to be placed in the thumbnail.}

\item{href}{URL to redirect when media or title is clicked.}
}
\description{
Bootstrap thumbnail
}
\examples{
if (interactive()) {



}
}
