% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{filter_cards_groups}
\alias{filter_cards_groups}
\title{Filter a Shuffle grid layout by groups}
\usage{
filter_cards_groups(session, shuffleId, groups)
}
\arguments{
\item{session}{The \code{session} object passed to function given to shinyServer.}

\item{shuffleId}{The id of the shuffle container.}

\item{groups}{Groups (defined in \code{shuffle_card}) you want to display.}
}
\description{
Filter a Shuffle grid layout by groups
}
\examples{

# TODO
}
