% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_shuffle}
\alias{update_shuffle}
\title{Update Shuffle Instance}
\usage{
update_shuffle(session, shuffleId)
}
\arguments{
\item{session}{The \code{session} object passed to function given to shinyServer.}

\item{shuffleId}{The id of the shuffle container.}
}
\description{
Update Shuffle Instance
}
\examples{
# TODO
}
